pyproject.toml
[project]
name = "slap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
package-mode = "false"
dependencies = [
    "matplotlib>=3.10.8",
    "mediapipe>=0.10.32",
    "opencv-python>=4.13.0.92",
    "scikit-learn>=1.8.0",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "turso-python (>=1.8,<2.0)",
]
